Extract asset cloning to separate class.
