+++ demo/src/main/java/com/novoda/storagepathfinder/demo/LandingActivity.java
@@ -17,27 +17,21 @@ import com.novoda.storagepathfinder.DeviceFeatures;
 import com.novoda.storagepathfinder.DeviceStorageRoot;
 import com.novoda.storagepathfinder.FileSystem;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.util.List;
 
 public class LandingActivity extends AppCompatActivity {
 
     private static final String ASSET_NAME = "10MB.zip";
-    private static final int BUFFER_SIZE = 1024;
-    private static final byte[] BUFFER = new byte[BUFFER_SIZE];
 
-    private AssetManager assetManager;
+    private AssetCloner assetCloner;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_landing);
 
-        assetManager = getApplicationContext().getAssets();
+        AssetManager assetManager = getApplicationContext().getAssets();
+        assetCloner = new AssetCloner(assetManager);
 
         CommonDirectories commonDirectories = new AndroidApplicationDirectories(getApplicationContext());
         FileSystem fileSystem = new AndroidFileSystem(commonDirectories);
@@ -65,49 +59,8 @@ public class LandingActivity extends AppCompatActivity {
     }
 
     private final DeviceStorageRootViewHolder.Listener onAddFileClicked = deviceStorageRoot -> {
-        File outputFile = new File(deviceStorageRoot.getAbsolutePath() + "/" + ASSET_NAME);
-        createFileIfDoesNotExist(outputFile);
-        copyAssetToFile(ASSET_NAME, outputFile);
+        String pathToCloneTo = deviceStorageRoot.getAbsolutePath() + "/" + ASSET_NAME;
+        assetCloner.cloneAsset(ASSET_NAME, pathToCloneTo);
     };
 
-    private void createFileIfDoesNotExist(File outputFile) {
-        boolean parentPathDoesNotExist = !outputFile.getParentFile().exists();
-        if (parentPathDoesNotExist) {
-            Log.w(getClass().getSimpleName(), String.format("path: %s doesn't exist, creating parent directories...", outputFile.getAbsolutePath()));
-            parentPathDoesNotExist = !outputFile.getParentFile().mkdirs();
-        }
-
-        if (parentPathDoesNotExist) {
-            throw new IllegalArgumentException("Unable to create path: " + outputFile.getParentFile().getAbsolutePath());
-        }
-    }
-
-    private void copyAssetToFile(String assetName, File outputFile) {
-        InputStream inputStream = null;
-        OutputStream myOutput = null;
-        int length;
-        try {
-            inputStream = assetManager.open(assetName);
-            myOutput = new FileOutputStream(outputFile, true);
-            while ((length = inputStream.read(BUFFER)) > 0) {
-                myOutput.write(BUFFER, 0, length);
-            }
-            Log.d(getClass().getSimpleName(), "Copied asset: " + assetName);
-        } catch (IOException e) {
-            Log.e(getClass().getSimpleName(), "Failed to copy asset: " + assetName, e);
-        } finally {
-            try {
-                if (myOutput != null) {
-                    myOutput.close();
-                    myOutput.flush();
-                }
-
-                if (inputStream != null) {
-                    inputStream.close();
-                }
-            } catch (IOException e) {
-                Log.e(getClass().getSimpleName(), "Failed to close streams.", e);
-            }
-        }
-    }
 }
